בלגןןןןן

+++ BLNEW/IBL.cs
@@ -17,7 +17,7 @@ namespace IBL
         #nullable disable
         //צריך לחשוב האם כדי להפוך את הבאים גם לצורה של השאר של קבלת ישויות 
         void DroneToCharge(int id);
-        void DroneOutCharge(int id, int time);
+        void DroneOutCharge(int id, double time);
         void AttacheDrone(int id);
         void PickUpParcel(int id);
         void Parceldelivery(int id);
